+++ DNN Platform/HttpModules/DotNetNuke.HttpModules.csproj
@@ -94,6 +94,7 @@
     <Compile Include="..\..\SolutionInfo.cs">
       <Link>SolutionInfo.cs</Link>
     </Compile>
+    <Compile Include="Compression\Config\Algorithms.cs" />
     <Compile Include="DependencyInjection\ServiceRequestScopeModule.cs" />
     <Compile Include="Localization\LocalizationModule.cs" />
     <Compile Include="MobileRedirect\MobileRedirectModule.cs" />
@@ -129,7 +130,6 @@
     <Compile Include="Analytics\Config\AnalyticsEngine.cs" />
     <Compile Include="Analytics\Config\AnalyticsEngineCollection.cs" />
     <Compile Include="Compression\CompressionModule.cs" />
-    <Compile Include="Compression\Config\Enums.cs" />
     <Compile Include="Compression\Config\Settings.cs" />
     <Compile Include="Compression\Filters\CompressingFilter.cs" />
     <Compile Include="Compression\Filters\DeflateFilter.cs" />
